Move `.only` and `debugger;` checks to pre-push
