+++ package.json
@@ -67,6 +67,7 @@
     "ci:test:plugin": "npx lerna run test --scope @dendronhq/plugin-core --stream",
     "watch": "npx lerna exec --scope $pkg -- tsc -p tsconfig.build.json --watch",
     "hooks:pre-commit": "node ./hooks/pre-commit.js",
+    "hooks:pre-push": "node ./hooks/pre-push.js",
     "gen:meta": "node ./bootstrap/scripts/genMeta.js",
     "gen:links": "node ./bootstrap/scripts/linkAll.js",
     "backend:updateDendronDeps": "node bootstrap/backend/updateDendronhqDeps.js"
@@ -78,7 +79,8 @@
   "useWorkspaces": false,
   "husky": {
     "hooks": {
-      "pre-commit": "lint-staged -v -d && yarn hooks:pre-commit"
+      "pre-commit": "lint-staged -v -d && yarn hooks:pre-commit",
+      "pre-push": "yarn hooks:pre-push"
     }
   },
   "lint-staged": {
