+++ packages/engine-test-utils/src/__tests__/engine-server/topics/hooks.spec.ts
@@ -222,7 +222,6 @@ describe("engine", async () => {
           },
           engine,
         }));
-        debugger;
         await expectNote({ note, match: ["daily body hello"] });
       },
       {
